+++ Simulation_Snooker/Ball.cpp
@@ -18,8 +18,8 @@ using namespace Engine::Graphics;
  *               ball)
  */
 Ball::Ball(const Vector2 &pos, int points)
-    : SphericalEntity(pos, RADIUS, false, 0.99f, 0.005f)
-    , RenderableObject(Mesh::GenerateDisc2D(RADIUS), NULL)
+    : SphericalEntity(pos, (float) RADIUS, false, 0.99f, 0.005f)
+    , RenderableObject(Mesh::GenerateDisc2D((float) RADIUS), NULL)
     , m_points(points)
 {
   ShaderProgram *sp = new ShaderProgram();
